+++ crates/polars-time/src/lib.rs
@@ -1,11 +1,17 @@
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#[cfg(feature = "timezones")]
 mod base_utc_offset;
 pub mod chunkedarray;
 mod date_range;
+#[cfg(feature = "timezones")]
 mod dst_offset;
 mod group_by;
+#[cfg(feature = "month_end")]
 mod month_end;
+#[cfg(feature = "month_start")]
 mod month_start;
+#[cfg(feature = "offset_by")]
+mod offset_by;
 pub mod prelude;
 mod round;
 pub mod series;
@@ -21,8 +27,12 @@ pub use date_range::*;
 pub use dst_offset::*;
 #[cfg(any(feature = "dtype-date", feature = "dtype-datetime"))]
 pub use group_by::dynamic::*;
+#[cfg(feature = "month_end")]
 pub use month_end::*;
+#[cfg(feature = "month_start")]
 pub use month_start::*;
+#[cfg(feature = "offset_by")]
+pub use offset_by::*;
 pub use round::*;
 pub use truncate::*;
 pub use upsample::*;
