+++ crates/polars/Cargo.toml
@@ -143,7 +143,9 @@ cum_agg = ["polars-ops/cum_agg", "polars-lazy?/cum_agg"]
 cumulative_eval = ["polars-lazy?/cumulative_eval"]
 cutqcut = ["polars-lazy?/cutqcut"]
 dataframe_arithmetic = ["polars-core/dataframe_arithmetic"]
-date_offset = ["polars-lazy?/date_offset"]
+month_start = ["polars-lazy?/month_start"]
+month_end = ["polars-lazy?/month_end"]
+offset_by = ["polars-lazy?/offset_by"]
 decompress = ["polars-io/decompress"]
 decompress-fast = ["polars-io/decompress-fast"]
 describe = ["polars-core/describe"]
