+++ py-polars/Cargo.toml
@@ -44,7 +44,9 @@ features = [
   "cum_agg",
   "cumulative_eval",
   "dataframe_arithmetic",
-  "date_offset",
+  "month_start",
+  "month_end",
+  "offset_by",
   "diagonal_concat",
   "diff",
   "dot_diagram",
